Move implementation files into lib/ directory
Now that `test_helper.rb` is responsible for putting the implementation
files on the load path for all tests, we only need to update which
directory it’s pointing at (versus changing a `require_relative` in
every test) to keep the tests working after the files move.

We do still need to update the test (and documentation) for
`ruby_hack_assembler.rb` because that gets run via the shell, which
doesn’t know about Ruby’s `$LOAD_PATH`.