Move test files into test/ directory
Again, we now only need to replace `"./lib"` with `"../lib"` in
`test_helper.rb` to support this change.

We did have to update the CI job to look for the tests in
`./test/*_test.rb` instead of `./*_test.rb` because it has no other way
of knowing where they are.