Make ruby_hack_assember.rb directly executable
Rather than expecting the assembler to be executed with `ruby`, we can
give it executable permissions (so the shell will run it directly) and
put a shebang line [0] at the start (so the shell knows to use Ruby to
run it). This makes it easier for a user to run.

[0] https://devcenter.heroku.com/articles/ruby-binstub-shebang