Rename assembler executable to just `assembler`
Now that this lives in `bin/` rather than `lib/` it doesn’t need a
unique name to distinguish it from the file which defines the
`Assembler` class, so we can give it a shorter and more convenient name.

The same goes for its integration test: now that it’s in
`test/integration` we can call it `assembler_test.rb` without it
conflicting with the existing `test/unit/assembler_test.rb`. But we do
need to give the integration test’s _class_ a unique name so that it can
be loaded at the same time as the `AssemblerTest` unit test without
accidentally reusing the same test class.