Add a Gemfile.lock
This isn’t strictly required, but it makes it easier to ensure that the
same versions of all dependencies are used everywhere.

I generated this lockfile on an Intel Mac (`x86_64-darwin-20`), but used
`bundle lock --add-platform x86_64-linux` [0] to add Intel Linux as a
platform so that `bundle install` works on GitHub Actions too.

[0] https://bundler.io/v2.2/bundle_lock.html#SUPPORTING-OTHER-PLATFORMS